move parseBitmap and getBuffer to same file since they share dependancies
